src/jd-generator/requirements.txt
openai-agents
langgraph
langchain
langchain_openai
langchain_core
langgraph-checkpoint-redis
PyJWT
